+++ src/main/java/org/server/remoteclass/controller/AuthController.java
@@ -5,6 +5,7 @@ import org.server.remoteclass.dto.LoginDto;
 import org.server.remoteclass.dto.TokenDto;
 import org.server.remoteclass.dto.TokenRequestDto;
 import org.server.remoteclass.dto.UserDto;
+import org.server.remoteclass.service.AuthService;
 import org.server.remoteclass.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,25 +18,25 @@ import javax.validation.Valid;
 @Slf4j
 @RestController
 public class AuthController {
-    private final UserService userService;
+    private final AuthService authService;
 
-    public AuthController(UserService userService){
-        this.userService = userService;
+    public AuthController(AuthService authService){
+        this.authService = authService;
     }
 
     @PostMapping("/signup")
     public ResponseEntity<UserDto> signup(@Valid @RequestBody UserDto userDto){
-        return ResponseEntity.status(HttpStatus.CREATED).body(userService.signup(userDto));
+        return ResponseEntity.status(HttpStatus.CREATED).body(authService.signup(userDto));
     }
 
     @PostMapping("/login")
     public ResponseEntity<TokenDto> authorize(@Valid @RequestBody LoginDto loginDto){
-        return ResponseEntity.ok(userService.login(loginDto));
+        return ResponseEntity.ok(authService.login(loginDto));
     }
 
     @PostMapping("/reissue")
     public ResponseEntity<TokenDto> reissue(@RequestBody TokenRequestDto tokenRequestDto){
-        return ResponseEntity.ok(userService.reissue(tokenRequestDto));
+        return ResponseEntity.ok(authService.reissue(tokenRequestDto));
     }
 
 }
