+++ src/main/java/org/server/remoteclass/service/CustomUserDetailsService.java
@@ -1,0 +1,42 @@
+package org.server.remoteclass.service;
+
+import lombok.RequiredArgsConstructor;
+import org.server.remoteclass.entity.UserEntity;
+import org.server.remoteclass.jpa.UserRepository;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Optional;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    @Transactional(readOnly = true)
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return userRepository.findByEmail(email)
+                .map(this::createUserDetails)
+                .orElseThrow(() -> new UsernameNotFoundException(email + "이 데이터베이스에 존재하지 않습니다."));
+    }
+
+    // DB에 User 값이 존재한다면 UserDetails 객체로 만들어서 리턴
+    private UserDetails createUserDetails(UserEntity user){
+        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(user.getAuthority().toString());
+        return new User(
+                String.valueOf(user.getEmail()),
+                        user.getPassword(),
+                        Collections.singleton(grantedAuthority)
+        );
+    }
+}
